Compatibility fix: when parsing html variables with <img tags, itexts… (#525)
* Compatibility fix: when parsing html variables with <img tags, itextsharp 4 creates (using HtmlWorker.ParseToList) a list of Paragraphs with chunks inside (and Image) which then are displayed as blocks in the PDF.
On the other hand, itextsharp 5, for the same case creates one Paragraph with many chunks that have the new property InLine=true. Those are displayed inline in the final PDF.
Besides, there is a difference in the space at the top, before the first image in an HTML, itextsharp 4 inserts a separation before the first image, so Leading is now set to 0 as a workaround.
The compatibliity fix applies to HTMl containing only image elements. When using PDfPTable, the elements that overflow the container are not clipped in contrast to the method using Columns.

* Replace itextsharp.dll 4.0.2.0 by itextsharp.dll 4.1.6.0 + AllowPartiallyTrustedCallers to allow execution on medium trust environments.
https://github.com/schourode/iTextSharp-LGPL

* Remove unused project.